payments/src/events/order-created-listener.ts
import {
  Listener,
  OrderCreatedEvent,
  Subjects,
} from '@sayinmehmet-ticketing/common';
import { Message } from 'node-nats-streaming';
import { queueGroupName } from './queue-group-name';
import { Order } from '../models/orders';

class OrderCreatedListener extends Listener<OrderCreatedEvent> {
  readonly subject = Subjects.OrderCreated;
  queueGroupName = queueGroupName;

  onMessage(data: OrderCreatedEvent['data'], msg: Message) {
    const order = Order.build({
      id: data.id,
      price: data.ticket.price,
      status: data.status,
      userId: data.userId,
      version: data.version,
    });

    order.save();

    msg.ack();
  }
}

export { OrderCreatedListener };
